refactor(card): move F0CardRow to Experimental section
Relocate F0CardRow into `src/experimental/F0CardRow` so it surfaces under the
"Experimental" Storybook section (with the 🚧 maturity tag) instead of
"Components". F0CardRow shares primitives with F0Card (CardActions, CardAlert,
CardAvatar, types), so those stay in `components/F0Card`; only F0CardRow and its
exclusive CardRowActions move, importing the shared pieces via `@/` aliases.

- `components/F0Card/index.tsx` keeps a `@deprecated` re-export to the new path
  for backward compatibility; `experimental/exports.ts` adds the canonical export.
- `F0HILActionConfirmation` (the one internal consumer) now imports F0CardRow
  from `@/experimental/F0CardRow`.
- Story tag flipped stable → experimental to match the new section.
