+++ packages/react/src/components/F0Card/index.tsx
@@ -1,2 +1,3 @@
 export * from "./F0Card"
-export * from "./F0CardRow"
+/** @deprecated F0CardRow has moved to @/experimental/F0CardRow. Import from there instead. */
+export * from "../../experimental/F0CardRow"
