+++ packages/react/src/experimental/F0CardRow/F0CardRow.tsx
@@ -13,9 +13,17 @@ import {
   type CardPrimaryAction,
   type CardSecondaryAction,
   type CardSecondaryLink,
-} from "./components/CardActions"
-import { CardAlertWrapper, alertBorderColor } from "./components/CardAlert"
-import { CardAvatar, type CardAvatarVariant } from "./components/CardAvatar"
+} from "@/components/F0Card/components/CardActions"
+import {
+  CardAlertWrapper,
+  alertBorderColor,
+} from "@/components/F0Card/components/CardAlert"
+import {
+  CardAvatar,
+  type CardAvatarVariant,
+} from "@/components/F0Card/components/CardAvatar"
+import { type CardAlertProps } from "@/components/F0Card/types"
+
 import {
   CardRowActions,
   type CardRowConfirmAction,
@@ -23,7 +31,6 @@ import {
   type CardRowStatus,
   cardRowClassName,
 } from "./components/CardRowActions"
-import { type CardAlertProps } from "./types"
 
 export interface F0CardRowProps {
   /**
