+++ packages/react/src/experimental/exports.ts
@@ -23,6 +23,7 @@ export * from "../kits/Charts/exports"
  * @deprecated F0ActionBar has moved to @/components/F0ActionBar. Import from there instead.
  */
 export * from "../components/F0ActionBar"
+export * from "./F0CardRow"
 export * from "./F0VersionHistory"
 export * from "./Forms/exports"
 export * from "./Information/exports"
